add comments to diagnostic_builder.rs
