+++ kclvm/compiler_base/Cargo.toml
@@ -6,4 +6,4 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-
+compiler-base-diagnostic = {path = "./diagnostic", version = "0.1.0"}
