+++ kclvm/compiler_base/error/src/lib.rs
@@ -1,5 +1,7 @@
+use std::panic;
+
 use compiler_base_diagnostic::{
-    pendant::*, Diagnostic, DiagnosticBuilder, Message, Position, Sentence,
+    pendant::*, Diagnostic, DiagnosticBuilder, Message, Position, Sentence, emitter::{Emitter, EmitterWriter},
 };
 use compiler_base_macros::DiagnosticBuilderMacro;
 
@@ -12,11 +14,13 @@ mod tests;
     title,
     msg = "For more information about this error, try `rustc --explain E0999`."
 )]
-#[error(title, msg = "oh no! this is an error!", code = "E012")] // 目前在这里多写没人管，少写不行，因为是从一个list里面往出找。
+#[error(title, msg = "oh no! this is an error!", code = "E012")]
 #[help(title, code = "E000", msg = "I need help !")]
 #[error(msg = "oh no! this is an error!")]
 pub struct ThisIsAnErr {
     #[position(msg = "err position")]
     pub pos: Position,
 }
+
+
