+++ kclvm/compiler_base/macros/Cargo.toml
@@ -9,6 +9,7 @@ edition = "2021"
 proc-macro = true
 
 [dependencies]
+compiler-base = {path="../", version = "0.1.0"}
 compiler-base-style = {path="../style", version="0.1.0"}
 compiler-base-diagnostic = {path = "../diagnostic", version = "0.1.0"}
 synstructure = "0.12.1"
